+++ CHANGELOG.md
@@ -13,6 +13,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+ - Update `UserActionData` to embed data about the targeted element.
+
 ### Deprecated
 
 ### Removed